# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posts/{postId}/comments": {
      "get": {
        "operationId": "PostsController_getComments",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from previous page",
            "schema": {
              "example": "MjAyNS0xMi0xOVQxMDoxMjowMC4wMDBafDEyMy1wb3N0LWlk",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment list",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommentsResponseDto"
                    }
                  ],
                  "example": {
                    "items": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": ["posts"]
      },
      "post": {
        "operationId": "PostsController_createComment",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommentDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": ["posts"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/ai/rewrite": {
      "post": {
        "operationId": "AiController_rewrite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiRewriteRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rewrite post content using an AI provider",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/AiRewriteResponseDto"
                    }
                  ],
                  "example": {
                    "content": "Shipping a fresh batch of ideas after today's research sprint — distilled and ready to share.",
                    "provider": "mock"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": ["ai"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Isntgram API",
    "description": "Isntgram API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PostAuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b6cf7a42-3f7a-4a7b-97a2-6c8a13d7b56e"
          },
          "username": {
            "type": "string",
            "example": "ava"
          },
          "fullName": {
            "type": "string",
            "example": "Ava Thompson"
          },
          "profilePictureUrl": {
            "type": "string",
            "example": "https://cdn.isntgram.ai/avatars/ava.jpg"
          }
        },
        "required": ["id", "username", "fullName"]
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "66efb64c-702f-46f5-9b44-7c9f2b0a0d61"
          },
          "postId": {
            "type": "string",
            "example": "7c1a2f04-1cf3-4c75-b42f-8f6c0f3e4d5a"
          },
          "content": {
            "type": "string",
            "example": "Great point — I ran into the same thing last week."
          },
          "likeCount": {
            "type": "number",
            "example": 12
          },
          "likedByViewer": {
            "type": "boolean",
            "example": false,
            "description": "Whether the authenticated viewer has liked this comment."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-19T10:12:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-12-19T10:12:00.000Z"
          },
          "author": {
            "$ref": "#/components/schemas/PostAuthorDto"
          }
        },
        "required": [
          "id",
          "postId",
          "content",
          "likeCount",
          "likedByViewer",
          "createdAt",
          "updatedAt",
          "author"
        ]
      },
...
      "CommentsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "description": "Opaque cursor from previous page",
            "example": "MjAyNS0xMi0xOVQxMDoxMjowMC4wMDBafDY2ZWZiNjRjLTcwMmYtNDZmNS05YjQ0LTdjOWYyYjBhMGQ2MQ=="
          }
        },
        "required": ["items"]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "example": "2025-12-19T12:34:56.789Z"
          },
          "path": {
            "type": "string",
            "example": "/api/posts/feed?limit=20"
          },
          "requestId": {
            "type": "string",
            "example": "req_9a6b3f0c-4a0a-4f11-9c61-2f6a3c5d1b2c"
          },
          "message": {
            "type": "string",
            "example": "Invalid cursor"
          },
          "errors": {
            "example": [
              "username must be longer than or equal to 3 characters"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string",
            "example": "BadRequestException"
          }
        },
        "required": ["statusCode", "timestamp", "path", "message", "error"]
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": ["content"]
      },
...
      "AiRewriteResponseDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Shipping a fresh batch of ideas after today's research sprint — distilled and ready to share."
          },
          "provider": {
            "type": "string",
            "enum": ["mock", "openai"]
          },
          "model": {
            "type": "string",
            "example": "gpt-4o-mini"
          }
        },
        "required": ["content", "provider"]
      },
      "AiRewriteRequestDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "example": "Shipping a fresh batch of ideas after today's research sprint."
          },
          "tone": {
            "type": "string",
            "enum": ["professional", "friendly", "concise"],
            "default": "professional"
          },
          "maxLength": {
            "type": "number",
            "description": "Maximum character length for the rewritten content",
            "default": 2000,
            "minimum": 1,
            "maximum": 2000
          }
        },
        "required": ["content"]
      }
    }
  }
}